+++ lib/kitchen/driver/oci.rb
@@ -68,6 +68,9 @@ module Kitchen
       default_config :use_token_auth, false
       default_config :shape_config, {}
       default_config :nsg_ids, []
+      default_config :all_plugins_disabled, false
+      default_config :management_disabled, false
+      default_config :monitoring_disabled, false
 
       # compute only configs
       default_config :setup_winrm, false