+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 1.18.0
 - feat: select image ocid for compute by display name with special sauce to select latest
+- feat: add `agent_config` property to compute launch details
 
 # 1.17.0
 - refactor: dry out instance classes and models